requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
Werkzeug==3.1.3
